Renamed folders to web-friendly & added index
Removed spaces from project folders and replaced with web-friendly names. Also added an index file at the top level of the project so you can easily open each project. To open the project file in the browser (eg. index-START.html), right-click the file in the expanded list and “Open in new tab”